% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateTextnet.R
\name{CreateTextnet}
\alias{CreateTextnet}
\title{Create Network Objects from text datasets}
\usage{
CreateTextnet(tidytextobject)
}
\arguments{
\item{tidytext}{Tidytext object created by PrepText or PrepTextSent}
}
\value{
Network object
}
\description{
This function calculates the term frequency inverse document frequency of each word, creates a weighted adjacency matrix where the rows and columns are the groups and the cells describe the sum of overlapping TFIDF between them; and c) creates an igraph object from this adjacency matrix. PrepTextSent objects will be multiplied by the sentiment and the returning cells are the overlapping TFIDF-sentiment products.
}
\examples{
\dontrun{CreateTextnet(tidytextobject)}

}
