% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepText.R
\name{PrepText}
\alias{PrepText}
\title{Prepare text for analysis}
\usage{
PrepText(
  textdata,
  groupvar,
  textvar,
  node_type = c("groups", "words"),
  tokenizer = c("words", "tweets"),
  pos = c("all", "nouns"),
  language = "english",
  remove_stop_words = FALSE,
  remove_numbers = NULL,
  compound_nouns = FALSE,
  udmodel_lang = NULL,
  ...
)
}
\arguments{
\item{groupvar}{Name of the column containing the grouping variable}

\item{textvar}{Name of the column containing the text}

\item{node_type}{Describes whether the user wants to create a network where the nodes are words within a document, or the nodes are the documents themselves.}

\item{tokenizer}{Describes whether the input data are tweets rather than regular text. If tokenizer = "tweets" then hashtag and mentions will be tokenized correctly, as well as the url removed.}

\item{pos}{If pos = "nouns" it will returns only nouns or proper nouns.}

\item{language}{}

\item{remove_stop_words}{Boolean. If true, it will automatically remove stopwords}

\item{remove_numbers}{Boolean. If true, it will automatically remove all numbers}

\item{compound_nouns}{Boolean. If true, it will parse noun compounds}

\item{udmodel_lang}{It provides the possibility to pass a preloaded udpipe language model}

\item{dataframe}{Dataframe}

\item{strip_punct}{Boolean. If true, it will automatically remove punctuation}
}
\value{
Dataframe summarized in tidytext format
}
\description{
The function creates a dataframe in summarized tidy text format (where each row of the dataset describes the prevalence of each word within the document).
}
\examples{
\dontrun{PrepText(textdata, groupvar, textvar)}

}
