% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextCommunities.R
\name{TextCommunities}
\alias{TextCommunities}
\title{Cluster network object}
\usage{
TextCommunities(text_network, alpha = 0.25)
}
\arguments{
\item{tidytext}{Tidytext object created by PrepText or PrepTextSent}
}
\value{
Network object
}
\description{
This function clusters the network objects based on the Louvain community detection algorithm
}
\examples{
\dontrun{TextCommunities(text_network, alpha =.50)}

}
